+++ Fdf_2/libft/src/ft_max.c
@@ -6,16 +6,16 @@
 /*   By: qbanet <qbanet@student.42perpignan.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/23 18:32:23 by qbanet            #+#    #+#             */
-/*   Updated: 2023/08/09 12:18:21 by qbanet           ###   ########.fr       */
+/*   Updated: 2023/08/10 14:23:48 by qbanet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_max(long long *tab, unsigned int len)
+int	ft_max(int *tab, unsigned int len)
 {
 	unsigned int	i;
-	long long		max;
+	int		max;
 
 	if (len == 0)
 		return (0);
