+++ Fdf_2/libft/src/ft_min.c
@@ -6,16 +6,16 @@
 /*   By: qbanet <qbanet@student.42perpignan.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/09 12:21:04 by qbanet            #+#    #+#             */
-/*   Updated: 2023/08/09 12:21:41 by qbanet           ###   ########.fr       */
+/*   Updated: 2023/08/10 14:23:39 by qbanet           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_min(long long *tab, unsigned int len)
+int	ft_min(int *tab, unsigned int len)
 {
 	unsigned int	i;
-	long long		min;
+	int		min;
 
 	if (len == 0)
 		return (0);
